Extract extensions to separate files
